feat(storage-purchases): handles no agreements case and extracts logic
